- Refactored the file structured, now the main UI and the functions are separated - Added two files: "utils" and "fileOperations". The first contains generic utils functions, the other files specfic functions
