+++ ChatSQL/code/fromJsonToEmb.py
@@ -27,7 +27,7 @@ def generate_emb_upsert_commands(json_file_path):
 if __name__ == "__main__":
     dir_path = os.path.dirname(os.path.realpath(__file__))
     json_file_path = os.path.join(dir_path, "..", "JSON")
-    fileDB = os.path.join(json_file_path,"movies.json") #va cambiata ogni volta che vuoi cambiare file
+    fileDB = os.path.join(json_file_path, "movies.json") #va cambiata ogni volta che vuoi cambiare file
 
     generated_commands = generate_emb_upsert_commands(fileDB)
     for command in generated_commands:
